+++ content.php
@@ -11,7 +11,7 @@
 		<?php if ( 'post' == get_post_type() ) : ?>
 		<div class="entry-meta">
 
-			<?php greenlake_posted_on(); ?>
+			<?php fremont_cut_posted_on(); ?>
 
 		</div><!-- .entry-meta -->
 		<?php endif; ?>
@@ -40,6 +40,6 @@
 	</div><!-- .entry-content -->
 
 	<footer class="entry-footer">
-		<?php greenlake_entry_footer(); ?>
+		<?php fremont_cut_entry_footer(); ?>
 	</footer><!-- .entry-footer -->
 </article><!-- #post-## -->