+++ header.php
@@ -12,13 +12,6 @@
 <meta charset="<?php bloginfo( 'charset' ); ?>">
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <link rel="pingback" href="<?php bloginfo( 'pingback_url' ); ?>">
-<?php
-echo "<meta name='description' content='";
-echo bloginfo('name');
-echo ": ";
-echo bloginfo( 'description' );
-echo "'>";
-?>
 <?php wp_head(); ?>
 </head>