+++ inc/custom-header.php
@@ -18,31 +18,31 @@
 /**
  * Set up the WordPress core custom header feature.
  *
- * @uses greenlake_header_style()
- * @uses greenlake_admin_header_style()
- * @uses greenlake_admin_header_image()
+ * @uses fremont_cut_header_style()
+ * @uses fremont_cut_admin_header_style()
+ * @uses fremont_cut_admin_header_image()
  */
-function greenlake_custom_header_setup() {
-	add_theme_support( 'custom-header', apply_filters( 'greenlake_custom_header_args', array(
+function fremont_cut_custom_header_setup() {
+	add_theme_support( 'custom-header', apply_filters( 'fremont_cut_custom_header_args', array(
 		'default-image'          => '',
 		'default-text-color'     => '000000',
 		'width'                  => 1400,
 		'height'                 => 250,
 		'flex-height'            => true,
-		'wp-head-callback'       => 'greenlake_header_style',
-		'admin-head-callback'    => 'greenlake_admin_header_style',
-		'admin-preview-callback' => 'greenlake_admin_header_image',
+		'wp-head-callback'       => 'fremont_cut_header_style',
+		'admin-head-callback'    => 'fremont_cut_admin_header_style',
+		'admin-preview-callback' => 'fremont_cut_admin_header_image',
 	) ) );
 }
-add_action( 'after_setup_theme', 'greenlake_custom_header_setup' );
+add_action( 'after_setup_theme', 'fremont_cut_custom_header_setup' );
 
-if ( ! function_exists( 'greenlake_header_style' ) ) :
+if ( ! function_exists( 'fremont_cut_header_style' ) ) :
 /**
  * Styles the header image and text displayed on the blog
  *
- * @see greenlake_custom_header_setup().
+ * @see fremont_cut_custom_header_setup().
  */
-function greenlake_header_style() {
+function fremont_cut_header_style() {
 	$header_text_color = get_header_textcolor();
 
 	// If no custom options for text are set, let's bail
@@ -75,15 +75,15 @@ function greenlake_header_style() {
 	</style>
 	<?php
 }
-endif; // greenlake_header_style
+endif; // fremont_cut_header_style
 
-if ( ! function_exists( 'greenlake_admin_header_style' ) ) :
+if ( ! function_exists( 'fremont_cut_admin_header_style' ) ) :
 /**
  * Styles the header image displayed on the Appearance > Header admin panel.
  *
- * @see greenlake_custom_header_setup().
+ * @see fremont_cut_custom_header_setup().
  */
-function greenlake_admin_header_style() {
+function fremont_cut_admin_header_style() {
 ?>
 	<style type="text/css">
 		.appearance_page_custom-header #headimg {
@@ -103,15 +103,15 @@ function greenlake_admin_header_style() {
 	</style>
 <?php
 }
-endif; // greenlake_admin_header_style
+endif; // fremont_cut_admin_header_style
 
-if ( ! function_exists( 'greenlake_admin_header_image' ) ) :
+if ( ! function_exists( 'fremont_cut_admin_header_image' ) ) :
 /**
  * Custom header image markup displayed on the Appearance > Header admin panel.
  *
- * @see greenlake_custom_header_setup().
+ * @see fremont_cut_custom_header_setup().
  */
-function greenlake_admin_header_image() {
+function fremont_cut_admin_header_image() {
 	$style = sprintf( ' style="color:#%s;"', get_header_textcolor() );
 ?>
 	<div id="headimg">
@@ -123,4 +123,4 @@ function greenlake_admin_header_image() {
 	</div>
 <?php
 }
-endif; // greenlake_admin_header_image
+endif; // fremont_cut_admin_header_image