+++ inc/customizer.php
@@ -10,17 +10,17 @@
  *
  * @param WP_Customize_Manager $wp_customize Theme Customizer object.
  */
-function greenlake_customize_register( $wp_customize ) {
+function fremont_cut_customize_register( $wp_customize ) {
 	$wp_customize->get_setting( 'blogname' )->transport         = 'postMessage';
 	$wp_customize->get_setting( 'blogdescription' )->transport  = 'postMessage';
 	$wp_customize->get_setting( 'header_textcolor' )->transport = 'postMessage';
 }
-add_action( 'customize_register', 'greenlake_customize_register' );
+add_action( 'customize_register', 'fremont_cut_customize_register' );
 
 /**
  * Binds JS handlers to make Theme Customizer preview reload changes asynchronously.
  */
-function greenlake_customize_preview_js() {
-	wp_enqueue_script( 'greenlake_customizer', get_template_directory_uri() . '/js/customizer.js', array( 'customize-preview' ), '20130508', true );
+function fremont_cut_customize_preview_js() {
+	wp_enqueue_script( 'fremont_cut_customizer', get_template_directory_uri() . '/js/customizer.js', array( 'customize-preview' ), '20130508', true );
 }
-add_action( 'customize_preview_init', 'greenlake_customize_preview_js' );
+add_action( 'customize_preview_init', 'fremont_cut_customize_preview_js' );