+++ inc/jetpack.php
@@ -10,10 +10,10 @@
  * Add theme support for Infinite Scroll.
  * See: http://jetpack.me/support/infinite-scroll/
  */
-function greenlake_jetpack_setup() {
+function fremont_cut_jetpack_setup() {
 	add_theme_support( 'infinite-scroll', array(
 		'container' => 'main',
 		'footer'    => 'page',
 	) );
 }
-add_action( 'after_setup_theme', 'greenlake_jetpack_setup' );
+add_action( 'after_setup_theme', 'fremont_cut_jetpack_setup' );