+++ inc/template-tags.php
@@ -65,11 +65,11 @@ function the_post_navigation() {
 }
 endif;
 
-if ( ! function_exists( 'greenlake_posted_on' ) ) :
+if ( ! function_exists( 'fremont_cut_posted_on' ) ) :
 /**
  * Prints HTML with meta information for the current post-date/time and author.
  */
-function greenlake_posted_on() {
+function fremont_cut_posted_on() {
 	$time_string = '<time class="entry-date published updated" datetime="%1$s">%2$s</time>';
 	if ( get_the_time( 'U' ) !== get_the_modified_time( 'U' ) ) {
 		$time_string = '<time class="entry-date published" datetime="%1$s">%2$s</time><time class="updated" datetime="%3$s">%4$s</time>';
@@ -97,28 +97,28 @@ function greenlake_posted_on() {
 }
 endif;
 
-if ( ! function_exists( 'greenlake_entry_footer' ) ) :
+if ( ! function_exists( 'fremont_cut_entry_footer' ) ) :
 /**
  * Prints HTML with meta information for the categories, tags and comments.
  */
-function greenlake_entry_footer() {
+function fremont_cut_entry_footer() {
 	// Hide category and tag text for pages.
 	if ( 'post' == get_post_type() ) {
 		/* translators: used between list items, there is a space after the comma */
 		echo "<span class='ct_section'>";
-		if ( get_the_category() && greenlake_categorized_blog() ) {
-		//	printf( '<span class="cat-links">' . __( 'Posted in %1$s ', 'fremont-cut' ) . '</span>', $categories_list );
-		echo "Category: ";
+		if ( get_the_category() && fremont_cut_categorized_blog() ) {
+	//	printf( '<span class="cat-links">' . __( 'Posted in %1$s ', 'fremont-cut' ) . '</span>', $categories_list );
+	//	echo "Category: ";
 		}
 
-		foreach((get_the_category()) as $category) {
-			echo "<span class='ct_2'>";
-  		  echo $category->cat_name . ' ';
-  		  echo "</span>";
-		}
+		// foreach((get_the_category()) as $category) {
+		// 	echo "<span class='ct_2'>";
+  // 		  echo $category->cat_name . ' ';
+  // 		  echo "</span>";
+		// }
 		$categories_list = get_the_category_list( __( ', ', 'fremont-cut' ) );
-		if ( $categories_list && greenlake_categorized_blog() ) {
-		//	printf( '<span class="cat-links">' . __( 'Posted in %1$s ', 'fremont-cut' ) . '</span>', $categories_list );
+		if ( $categories_list && fremont_cut_categorized_blog() ) {
+			printf( '<span class="cat-links">' . __( 'Posted in %1$s ', 'fremont-cut' ) . '</span>', $categories_list );
 		}
 		echo "</span>";
 
@@ -126,20 +126,21 @@ function greenlake_entry_footer() {
 		$tags_list = get_the_tag_list( '', __( ', ', 'fremont-cut' ) );
 		if ( $tags_list ) {
 		echo "<span class='tg_section'>";
-			echo "Tagged: ";
+//			echo "Tagged: ";
 //			echo get_the_tag_list();
 			$posttags = get_the_tags();
 if ($posttags) {
 
 	foreach($posttags as $tag) {
-		echo "<span class='tg_2'>";
-		echo $tag->name . ' ';
-		echo "</span>";
+//		echo "<span class='tg_2'>";
+//		echo $tag->name . ' ';
+//		echo "</span>";
 	}
 
 }
+		
+		printf( '<span class="tags-links">' . __( 'Tagged %1$s ', 'fremont-cut' ) . '</span>', $tags_list );
 		echo "</span>";
-		//	printf( '<span class="tags-links">' . __( 'Tagged %1$s ', 'fremont-cut' ) . '</span>', $tags_list );
 		}
 	}
 
@@ -252,8 +253,8 @@ endif;
  *
  * @return bool
  */
-function greenlake_categorized_blog() {
-	if ( false === ( $all_the_cool_cats = get_transient( 'greenlake_categories' ) ) ) {
+function fremont_cut_categorized_blog() {
+	if ( false === ( $all_the_cool_cats = get_transient( 'fremont_cut_categories' ) ) ) {
 		// Create an array of all the categories that are attached to posts.
 		$all_the_cool_cats = get_categories( array(
 			'fields'     => 'ids',
@@ -266,27 +267,27 @@ function greenlake_categorized_blog() {
 		// Count the number of categories that are attached to the posts.
 		$all_the_cool_cats = count( $all_the_cool_cats );
 
-		set_transient( 'greenlake_categories', $all_the_cool_cats );
+		set_transient( 'fremont_cut_categories', $all_the_cool_cats );
 	}
 
 	if ( $all_the_cool_cats > 1 ) {
-		// This blog has more than 1 category so greenlake_categorized_blog should return true.
+		// This blog has more than 1 category so fremont_cut_categorized_blog should return true.
 		return true;
 	} else {
-		// This blog has only 1 category so greenlake_categorized_blog should return false.
+		// This blog has only 1 category so fremont_cut_categorized_blog should return false.
 		return false;
 	}
 }
 
 /**
- * Flush out the transients used in greenlake_categorized_blog.
+ * Flush out the transients used in fremont_cut_categorized_blog.
  */
-function greenlake_category_transient_flusher() {
+function fremont_cut_category_transient_flusher() {
 	if ( defined( 'DOING_AUTOSAVE' ) && DOING_AUTOSAVE ) {
 		return;
 	}
 	// Like, beat it. Dig?
-	delete_transient( 'greenlake_categories' );
+	delete_transient( 'fremont_cut_categories' );
 }
-add_action( 'edit_category', 'greenlake_category_transient_flusher' );
-add_action( 'save_post',     'greenlake_category_transient_flusher' );
+add_action( 'edit_category', 'fremont_cut_category_transient_flusher' );
+add_action( 'save_post',     'fremont_cut_category_transient_flusher' );